+++ modules/sql-engine/src/test/java/org/apache/ignite/internal/sql/engine/planner/datatypes/utils/Types.java
@@ -51,7 +51,9 @@ public final class Types {
 
     public static final NativeType DECIMAL_5_0 = NativeTypes.decimalOf(5, 0);
     public static final NativeType DECIMAL_6_1 = NativeTypes.decimalOf(6, 1);
+    public static final NativeType DECIMAL_6_4 = NativeTypes.decimalOf(6, 4);
     public static final NativeType DECIMAL_8_3 = NativeTypes.decimalOf(8, 3);
+    public static final NativeType DECIMAL_8_6 = NativeTypes.decimalOf(8, 6);
     public static final NativeType DECIMAL_21_16 = NativeTypes.decimalOf(21, 16);
 
     public static final NativeType DECIMAL_14_7 = NativeTypes.decimalOf(14, 7);
